# docs/swagger.yaml
basePath: /v1
definitions:
  dto.Pagination:
    properties:
      total:
        type: integer
    type: object
  dto.PaginationResponse:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
    type: object
  model.LoginRequest:
    properties:
      password:
        example: password
        type: string
      username:
        example: admin
        type: string
    type: object
  model.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
  model.UserCreateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role_id:
        type: string
      status:
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - phone_number
    - status
    - username
    type: object
  model.UserUpdateRequest:
    properties:
      email:
        type: string
      gender:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role_id:
        type: string
      status:
        type: string
      username:
        type: string
    required:
    - email
    - gender
    - name
    - password
    - phone_number
    - status
    - username
    type: object
host: localhost:8000
info:
  contact:
    email: arifsuganda.su@gmail.com
    name: Arif Suganda
    url: nevangga.com
  description: Aplikasi pengolahan pelanggan, transaksi dan aset perusahaan.
  license:
    name: EULA
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                Data:
                  $ref: '#/definitions/model.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Login
      tags:
      - Auth
  /users:
    get:
      consumes:
      - application/json
      description: List user with filters
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: username
        in: query
        name: username
        type: string
      - description: email
        in: query
        name: email
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: per_page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/dto.PaginationResponse'
                  - properties:
                      Items:
                        items:
                          $ref: '#/definitions/model.User'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User Create Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
      summary: Delete a user
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Update a user with the provided information
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User Update Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                message:
                  type: string
              type: object
      summary: Update a user
      tags:
      - Users
swagger: "2.0"